+++ app/helpers/dorsale/all_helpers.rb
@@ -13,6 +13,7 @@ module Dorsale::AllHelpers
   include ::Dorsale::BootstrapHelper
 
   include ::Dorsale::Alexandrie::AttachmentsHelper
+  include ::Dorsale::BillingMachine::ApplicationHelper
   include ::Dorsale::Flyboy::ApplicationHelper
   include ::Dorsale::ExpenseGun::ApplicationHelper
   include ::Dorsale::UsersHelper