+++ features/step_definitions/customer_vault_links_steps.rb
@@ -1,10 +1,12 @@
 
 Given(/^an existing corporation$/) do
-  @corporation= create(:customer_vault_corporation)
+  @corporation = create(:customer_vault_corporation)
+  @person = @corporation
 end
 
 Given(/^an existing indidual$/) do
   @individual = create(:customer_vault_individual)
+  @person = @individual
 end
 
 Given(/^an existing link$/) do
@@ -19,12 +21,12 @@ When(/^I navigate to the link section of the indidual details$/) do
 end
 
 When(/^I add a new link to the corporation$/) do
-  find(".tab-content [href$='links/new']").click
+  find("#context-main [href$='links/new']").click
   expect(current_path).to eq dorsale.new_customer_vault_corporation_link_path(@corporation)
 end
 
 When(/^I add a new link to the individual$/) do
-  find(".tab-content [href$='links/new']").click
+  find("#context-main [href$='links/new']").click
   expect(current_path).to eq dorsale.new_customer_vault_individual_link_path(@individual)
 end